Audio2Score-backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Audio processing
basic-pitch==0.2.5
tensorflow==2.13.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
requests==2.31.0
